refactor: procedure storage (#279)
## Rationale
The current design of the procedure storage module is unreasonable. The
data will be stored permanently and will not be cleared. This will cause
the storage space of `ETCD` to be filled up quickly. We need to redesign
the interface of procedure storage.

## Detailed Changes
* Add `ttl` config for `CreateOrUpdate`.
* Add `Delete` interface, it used to delete history procedure data and
current procedure data.

## Test Plan
Pass all unit tests and integration tests.